Addressed review comments to rename the classes.
