+++ core/src/main/scala/kafka/server/checkpoints/LeaderEpochCheckpointFile.scala
@@ -18,7 +18,7 @@ package kafka.server.checkpoints
 
 import kafka.server.LogDirFailureChannel
 import kafka.server.epoch.EpochEntry
-import org.apache.kafka.server.common.SnapshotFile.EntryFormatter
+import org.apache.kafka.server.common.CheckpointFile.EntryFormatter
 
 import java.io._
 import java.util.Optional
@@ -66,7 +66,7 @@ object LeaderEpochCheckpointFile {
 class LeaderEpochCheckpointFile(val file: File, logDirFailureChannel: LogDirFailureChannel = null) extends LeaderEpochCheckpoint {
   import LeaderEpochCheckpointFile._
 
-  val checkpoint = new CheckpointFile[EpochEntry](file, CurrentVersion, Formatter, logDirFailureChannel, file.getParentFile.getParent)
+  val checkpoint = new CheckpointFileWithFailureHandler[EpochEntry](file, CurrentVersion, Formatter, logDirFailureChannel, file.getParentFile.getParent)
 
   def write(epochs: Iterable[EpochEntry]): Unit = checkpoint.write(epochs)
