+++ core/src/main/scala/kafka/server/checkpoints/OffsetCheckpointFile.scala
@@ -19,7 +19,7 @@ package kafka.server.checkpoints
 import kafka.server.LogDirFailureChannel
 import kafka.server.epoch.EpochEntry
 import org.apache.kafka.common.TopicPartition
-import org.apache.kafka.server.common.SnapshotFile.EntryFormatter
+import org.apache.kafka.server.common.CheckpointFile.EntryFormatter
 
 import java.io._
 import java.util.Optional
@@ -62,7 +62,7 @@ trait OffsetCheckpoint {
  *  -----checkpoint file end----------
  */
 class OffsetCheckpointFile(val file: File, logDirFailureChannel: LogDirFailureChannel = null) {
-  val checkpoint = new CheckpointFile[(TopicPartition, Long)](file, OffsetCheckpointFile.CurrentVersion,
+  val checkpoint = new CheckpointFileWithFailureHandler[(TopicPartition, Long)](file, OffsetCheckpointFile.CurrentVersion,
     OffsetCheckpointFile.Formatter, logDirFailureChannel, file.getParent)
 
   def write(offsets: Map[TopicPartition, Long]): Unit = checkpoint.write(offsets)
